Move autosave thread functions to autosave.py, fix try block logic
