+++ softioc/autosave.py
@@ -4,8 +4,9 @@ import threading
 from datetime import datetime
 from pathlib import Path
 from numpy import ndarray
-
 from softioc.device_core import LookupRecordList
+import sys
+import atexit
 
 SAV_SUFFIX = "softsav"
 SAVB_SUFFIX = "softsavB"
@@ -32,22 +33,32 @@ def configure(directory=None, save_period=None, enabled=True, device=None):
         device: string name of the device prefix used for naming autosave files,
             automatically supplied by builder if not explicitly provided.
     '''
+    # if already set, do not overwrite save_period or directory
     Autosave.save_period = save_period or Autosave.save_period
     Autosave.enabled = enabled
+    if directory is not None:
+        Autosave.directory = Path(directory)
     if device is None:
         if Autosave.device_name is None:
             from .builder import GetRecordNames
             Autosave.device_name = GetRecordNames().prefix[0]
     else:
         Autosave.device_name = device
-    if directory is None and Autosave.directory is None:
-        raise RuntimeError(
-            "Autosave directory is not known, call "
-            "autosave.configure() with keyword argument "
-            "directory."
-        )
-    else:
-        Autosave.directory = Path(directory)
+
+
+def start_autosave_thread():
+    autosaver = Autosave()
+    worker = threading.Thread(
+        target=autosaver.loop,
+    )
+    worker.daemon = True
+    worker.start()
+    atexit.register(_shutdown_autosave_thread, autosaver, worker)
+
+
+def _shutdown_autosave_thread(autosaver, worker):
+    autosaver.stop()
+    worker.join()
 
 
 class Autosave:
@@ -66,22 +77,22 @@ class Autosave:
         yaml.add_representer(
             ndarray, _ndarray_representer, Dumper=yaml.Dumper
         )
+        if not self.device_name:
+            raise RuntimeError(
+                "Device name is not known to autosave thread, "
+                "call autosave.configure() with keyword argument device"
+            )
         if not self.directory:
             raise RuntimeError(
                 "Autosave directory is not known, call "
                 "autosave.configure() with keyword argument "
-                "directory."
+                "directory"
             )
-        if not self.device_name:
-            raise RuntimeError(
-                "Device name is not known to autosave thread, "
-                "call autosave.configure() with device keyword argument"
-            )
-        self._last_saved_time = datetime.now()
         if not self.directory.is_dir():
-            raise RuntimeError(
+            raise FileNotFoundError(
                 f"{self.directory} is not a valid autosave directory"
             )
+        self._last_saved_time = datetime.now()
         if self.backup_on_restart:
             self._backup_sav_file()
         self._pvs = {name: pv for name, pv in LookupRecordList() if pv.autosave}
@@ -92,7 +103,9 @@ class Autosave:
         if sav_path.is_file():
             shutil.copy2(sav_path, self._get_timestamped_backup_sav_path())
         else:
-            print(f"Could not back up autosave, {sav_path} is not a file")
+            sys.stderr.write(
+                f"Could not back up autosave, {sav_path} is not a file"
+            )
 
     def _get_timestamped_backup_sav_path(self):
         sav_path = self._get_current_sav_path()
@@ -116,27 +129,31 @@ class Autosave:
                 ]:
                     with open(path, "w") as f:
                         yaml.dump(state, f, indent=4)
-                self._last_saved_state = state.copy()
+                self._last_saved_state = state
                 self._last_saved_time = datetime.now()
         except Exception as e:
-            print(f"Could not save state to file: {e}")
+            sys.stderr.write(f"Could not save state to file: {e}")
 
     def _load(self, path=None):
         if not self.enabled:
-            print("Not loading from file as autosave adapter disabled")
+            sys.stdout.write(
+                "Not loading from file as autosave adapter disabled"
+            )
             return
         sav_path = path or self._get_current_sav_path()
         if not sav_path or not sav_path.is_file():
-            print(f"Could not load autosave values from file {sav_path}")
+            sys.stderr.write(
+                f"Could not load autosave values from file {sav_path}"
+            )
             return
         with open(sav_path, "r") as f:
             self._last_saved_state = yaml.full_load(f)
         for name, value in self._last_saved_state.items():
-            pv = self._pvs.get(name, None)
-            if not pv:
-                print(f"{name} is not a valid autosaved PV")
-                continue
-            pv.set(value)
+            try:
+                pv = self._pvs.get(name)
+                pv.set(value)
+            except Exception as e:
+                sys.stderr.write(f"Exception setting {name} to {value}: {e}")
 
     def stop(self):
         self._stop_event.set()
@@ -144,12 +161,12 @@ class Autosave:
     def loop(self):
         if not self.enabled:
             return
-        try:
-            while True:
+        while True:
+            try:
                 self._stop_event.wait(timeout=self.save_period)
                 if self._stop_event.is_set():  # Stop requested
                     return
                 else:  # No stop requested, we should save and continue
                     self._save()
-        except Exception as e:
-            print(f"Exception in autosave loop: {e}")
+            except Exception as e:
+                sys.stderr.write(f"Exception in autosave loop: {e}")
